requirements.txt
streamlit==1.41.1
langchain==0.3.16
langchain-community==0.3.16
faiss-cpu
sentence-transformers
huggingface-hub
ollama
pdfplumber
pypdf
numpy==1.26.4
pandas
torch
python-docx
rank-bm25
langchain-text-splitters
rank-bm25
sentence-transformers